+++ src/main/java/com/delegrego/exemplo_spring_boot_2/repo/FuncionarioRepository.java
@@ -1,6 +1,5 @@
 package com.delegrego.exemplo_spring_boot_2.repo;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
@@ -8,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.delegrego.exemplo_spring_boot_2.entity.FuncionarioEntity;
 
-public interface FuncionarioRepository extends JpaRepository<FuncionarioEntity, BigInteger> {
+public interface FuncionarioRepository extends JpaRepository<FuncionarioEntity, Long> {
 
 	// Derived queries
 
@@ -58,7 +57,7 @@ public interface FuncionarioRepository extends JpaRepository<FuncionarioEntity, 
 	 * @return Um boolean indicando se o cpf já existe para outro funcionário
 	 *
 	 */
-	boolean existsByCpfAndIdFuncionarioNot(String cpf, BigInteger id);
+	boolean existsByCpfAndIdFuncionarioNot(String cpf, Long id);
 
 	/**
 	 * Retorna um boolean indicando se o email já existe para outro funcionário
@@ -70,7 +69,7 @@ public interface FuncionarioRepository extends JpaRepository<FuncionarioEntity, 
 	 *         diferente do fornecido
 	 *
 	 */
-	boolean existsByEmailAndIdFuncionarioNot(String email, BigInteger id);
+	boolean existsByEmailAndIdFuncionarioNot(String email, Long id);
 
 	/**
 	 * Retorna um boolean indicando se existe algum funcionário associado ao
@@ -80,5 +79,5 @@ public interface FuncionarioRepository extends JpaRepository<FuncionarioEntity, 
 	 * @return Um boolean indicando se existe algum funcionário associado ao
 	 *         departamento fornecido
 	 */
-	boolean existsByDepartamentoIdDepartamento(BigInteger id);
+	boolean existsByDepartamentoIdDepartamento(Long id);
 }
